Add persistence for grocery items.
There is a known issue with synchronizition. Currently meals in the
future are always passed when looking for ingredients and meals in the
past are only passed once (if no ingredients have been stored yet). The
comparison to date needs to be the date the meal was created and not the
day the meal will happen. This is just a limitation of the api response
that I need to fix.